requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.1.2
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.13.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.7.0
zope.interface==8.0.1
